make plugin, and trigger methods a part of emitter
